pocciリポジトリの更新機能を追加

+++ bin/get-updated-images.sh
@@ -2,8 +2,8 @@
 set -e
 
 cd $(dirname $0)/js
-docker images | tail -n +2 | grep $1 "xpfriend/" | awk 'BEGIN{printf "["}NR>1{printf ", "}{printf "\"%s:%s\"",$1,$2}END{printf "]\n"}' >./version.json
-
 IMAGE=`docker images |grep workspace-nodejs | head -1 | awk '{printf "%s:%s",$1,$2}'`
-docker run --rm -v ${PWD}:/app -w /app ${IMAGE} npm install
-docker run --rm -v ${PWD}:/app -w /app ${IMAGE} node ./get-updated-images.js | tr -d "\r"
+if [ ! -d node_modules ]; then
+    docker run --rm -v ${PWD}:/app -w /app ${IMAGE} npm install
+fi
+docker run --rm -v ${PWD}:/app -w /app ${IMAGE} node ./get-updated-images.js $1 | tr -d "\r"
